+++ x/genutil/types/genesis_state.go
@@ -20,6 +20,10 @@ type GenesisState struct {
 
 // NewGenesisState creates a new GenesisState object
 func NewGenesisState(genTxs []json.RawMessage) GenesisState {
+	// Ensure genTxs is never nil, https://github.com/cosmos/cosmos-sdk/issues/5086
+	if len(genTxs) == 0 {
+		genTxs = make([]json.RawMessage, 0)
+	}
 	return GenesisState{
 		GenTxs: genTxs,
 	}